Added multiple ratings support, made code more modular - Genre WIP
Modularized some of the code and broke it up more instead of large functions. Also added support to show multiple ratings, the type of rating (critic or audience), and where the ratings originated from. Implemented more styling to output sent to Discord so it is easier to read.
Next to work on is 'genre'