Added more DB functionality/Checks for duplicate messages
Added more functions for DB tables. Config file has been edited to have an 'elapsed_time' that is 12 hours in milliseconds that will be a check for duplicate messages (this can be changed). Plex events now move into the table when posted, the time is recorded. When a new event is posted the table is check against the 'timestamp', and title to see if it has been posted within 12 hours, if so it will not send a message to Discord. If more then 12 hours have passed, the message can be sent again if needed.