discordMessageService.js and config
discordMessageService.js and config were updated. Config now reports '720' for 'elapsed_time' which is 12 hours. This is for readability. discordMessageService.js has const elapsed_time_in_ms, to take the elapsed_time from the config file, convert it to milliseconds for computer accuracy and than uses that to calculate the cooldown for duplicate title post a Discord message.