Fix - markupService.js episode or show styling
First episodes were not posting correctly and were 'undefined' for the heading. In markupService.js, a simple fix was to change the 'else if' to else for the 'type' check. If content is not explicitly marked as 'movie', it will now be treated as a tv posting for heading purposes.