Added logging for plexWebhooks, seasons for ep display, style/formatting
Removed the 'n/a' for genre when Episodes are posted. It originally said 'n/a' because tv shows did not have a genre. Now there will nothing displayed if there is no genre for a cleaner looking message. Added the season to be displayed in a discord message along time tv show titles, and episode title. Logging has been added to see webhooks. Services<logPlexDataService.js now has a logger function to create and print out logs for webhooks sent to view in logs<plexWebhook.log.